[Gold IV] Title: 그룹 나누기 (Subset), Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/16132. 그룹 나누기 （Subset）/그룹 나누기 （Subset）.cc
@@ -14,7 +14,7 @@ int main(void){
         return 0;
     }
     total /= 2;
-    vector<int> dp(total+1, 0);
+    vector<long long> dp(total+1, 0);
     dp[0] += 1;
     for (int i=1; i<=n; i++){
         for (int j=total; j>=i; j--){
